* Component renamed * Added encoding parameter * Other fixes
