+++ homeassistant/components/google_cloud/__init__.py
@@ -1,0 +1,1 @@
+"""The google_cloud component."""
