--- homeassistant/components/google_cloud_tts/__init__.py
@@ -1,1 +1,0 @@
-"""The google_cloud_tts component."""
